refactor(interpreter): split captures guards and dispatch
